+++ crates/gen/src/lib.rs
@@ -27,6 +27,7 @@ mod to_ident;
 mod vector2;
 mod vector3;
 mod vector4;
+mod guid;
 
 pub use gen::*;
 pub use helpers::*;
@@ -60,3 +61,4 @@ use timespan::*;
 use vector2::*;
 use vector3::*;
 use vector4::*;
+use guid::*;
